+++ bt2_idx.h
@@ -3587,7 +3587,7 @@ void Ebwt<index_t>::buildToDisk(
 						size_t idx = acc_szs.bsearchLoBound(saElt);
 						assert_lt(idx, acc_szs.size());
 						bool different = false;
-						for(size_t k = 0; k < kmer_size; k++) {
+						for(size_t k = 0; k < (size_t)kmer_size; k++) {
 							if((acc_szs[idx]-saElt) > k) {
 								uint8_t bp = s[saElt+k];
 								if(kmer[k] != bp || kmer_count[k] <= 0 || different) {
@@ -3790,7 +3790,7 @@ void Ebwt<index_t>::buildToDisk(
 	}
     
     if(kmer_size > 0) {
-      for(size_t k = 0; k < kmer_size; k++) {
+      for(size_t k = 0; k < (size_t)kmer_size; k++) {
         cerr << "Number of distinct " << k+1 << "-mers is " << kmer_count[k] << endl;
       }
     }
